Release tasks with Gradle
Fix version increment
Rename tasks
Declare task dependencies
